% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fledge-package.R
\docType{package}
\name{fledge-package}
\alias{fledge}
\alias{fledge-package}
\title{fledge: Smoother Change Tracking and Versioning for R Packages}
\description{
Streamlines the process of updating changelogs (NEWS.md) and versioning R packages developed in git repositories.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fledge.cynkra.com/}
  \item \url{https://github.com/cynkra/fledge}
  \item Report bugs at \url{https://github.com/cynkra/fledge/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{kirill@cynkra.com} (\href{https://orcid.org/0000-0002-1416-3412}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Patrick Schratz \email{patrick.schratz@gmail.com} (\href{https://orcid.org/0000-0003-0748-6624}{ORCID})
}

Other contributors:
\itemize{
  \item Maëlle Salmon \email{msmaellesalmon@gmail.com} (\href{https://orcid.org/0000-0002-2815-0399}{ORCID}) [contributor]
}

}
\keyword{internal}
