% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{release_deprecated}
\alias{release_deprecated}
\alias{init_release}
\alias{pre_release}
\title{Automating CRAN release}
\usage{
init_release(which = "next", force = FALSE)

pre_release(force = FALSE)
}
\arguments{
\item{which}{Component of the version number to update. Supported
values are
\itemize{
\item \code{"next"} (\code{"major"} if the current version is \code{x.99.99.9yz},
\code{"minor"} if the current version is \code{x.y.99.za},
\code{"patch"} otherwise),
\item \code{"patch"}
\item \code{"minor"},
\item \code{"major"}.
}}

\item{force}{Create branches and tags even if they exist.
Useful to recover from a previously failed attempt.}
}
\description{
\code{init_release()} and \code{pre_release()} are run
when a milestone in the development of a package is reached
and it is ready to be sent to CRAN.
}
\details{
\code{init_release()}:
\itemize{
\item Ensures that no modified files are in the git index.
\item Creates a release branch and bumps version to a non-development version which should be sent to CRAN.
\item Writes/updates \code{cran-comments.md} with useful information about the current
release process.
\item Prompts the user to run \code{urlchecker::url_update()}, \code{devtools::check_win_devel()},
and \code{rhub::check_for_cran()}.
}

\code{pre_release()}:
\itemize{
\item Opens a pull request for the release branch for final checks.
}
}
