% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_github_action_fledge.R
\name{use_github_action_pkgcheck}
\alias{use_github_action_pkgcheck}
\title{Use fledge GitHub Actions workflow}
\usage{
use_github_action_pkgcheck(save_as = NULL)
}
\arguments{
\item{save_as}{Name of the local workflow file. Defaults to \code{name} or
\code{fs::path_file(url)}. Do not specify any other part of the path; the parent
directory will always be \code{.github/workflows}, within the active project.}
}
\value{
Nothing, called for its side-effect.
}
\description{
Use fledge GitHub Actions workflow
}
