% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo-helper.R
\name{with_demo_project}
\alias{with_demo_project}
\alias{local_demo_project}
\title{Run code in temporary project}
\usage{
with_demo_project(code, dir = NULL, news = TRUE, quiet = FALSE)

local_demo_project(
  dir = NULL,
  news = TRUE,
  quiet = FALSE,
  .local_envir = parent.frame()
)
}
\arguments{
\item{code}{Code to run with temporary active project}

\item{dir}{Directory within which to create the mock package folder.}

\item{news}{If TRUE, create a NEWS.md file.}

\item{quiet}{Whether to show messages from usethis}

\item{.local_envir}{The environment to use for scoping. Defaults to current
execution environment.}
}
\value{
\code{with_demo_project()} returns the result of evaluating \code{code}.

\code{local_demo_project()} is called for its side effect and returns \code{NULL}, invisibly.
}
\description{
Run code in temporary project
}
\examples{
with_demo_project({
  # Add a new R file.
  usethis::use_r("cool-function", open = FALSE)
  # Pretend we added useful code inside it.
  # Track the new R file with Git.
  gert::git_add("R/cool-function.R")
  gert::git_commit("- Add cool function.")
  # Bump version with fledge.
  fledge::bump_version()
})
}
