% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-bump-version.R, R/bump-version.R
\name{bump_version}
\alias{bump_version}
\alias{bump_version_impl}
\title{Bump package version}
\usage{
bump_version(which = "dev")
}
\arguments{
\item{which}{Component of the version number to update. Supported
values are \code{"dev"} (default), \code{"patch"}, \code{"minor"} and \code{"major"}.}
}
\description{
Calls the following functions:

\enumerate{
\item \code{\link[=update_news]{update_news()}}
\item \code{\link[=update_version]{update_version()}}, using the \code{which} argument
\item Depending on the \code{which} argument:
\itemize{
\item If \code{"dev"}, \code{\link[=finalize_version]{finalize_version()}} with \code{push = FALSE}
\item Otherwise, \code{\link[=commit_version]{commit_version()}}.
}
}
}
