% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-bump-version.R
\name{bump_version}
\alias{bump_version}
\title{Bump package version}
\usage{
bump_version(which = "dev")
}
\arguments{
\item{which}{which}
}
\description{
Calls the following functions:

\enumerate{
\item \code{\link[=update_news]{update_news()}}
\item \code{\link[=update_version]{update_version()}}, using the \code{which} argument
\item \code{\link[=tag_version]{tag_version()}}
}
}
