% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-commit-version.R
\name{commit_version}
\alias{commit_version}
\title{Commits NEWS.md and DESCRIPTION to Git}
\usage{
commit_version()
}
\value{
Invisibly: \code{TRUE} if a previous commit for that version has been
amended, \code{FALSE} if not.
}
\description{
Commits changes to \code{NEWS.md} and \code{DESCRIPTION}, amending a previous commit
created by \pkg{fledge} if necessary.
}
