% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo-helper.R
\name{create_demo_project}
\alias{create_demo_project}
\title{Create example repo for fledge demos}
\usage{
create_demo_project(
  open = rlang::is_interactive(),
  pkg = "tea",
  name = "Kirill Müller",
  email = "mail@example.com",
  date = "2020-12-12",
  tempdir = withr::local_tempdir(pattern = "fledge", .local_envir = parent.frame(n =
    2)),
  news = FALSE
)
}
\arguments{
\item{open}{Whether to open the new project.}

\item{pkg}{Package name.}

\item{name}{Name for DESCRIPTION and git.}

\item{email}{Email for DESCRIPTION and git.}

\item{date}{String of time for DESCRIPTION and git.}

\item{tempdir}{Directory within which to create the mock package folder.}

\item{news}{If TRUE, create a NEWS.md file.}
}
\value{

}
\description{
Create example repo for fledge demos
}
