% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo-helper.R
\name{create_demo_project}
\alias{create_demo_project}
\title{Create example repo for fledge demos}
\usage{
create_demo_project(
  open = rlang::is_interactive(),
  name = "tea",
  maintainer = whoami::fullname(fallback = "Kirill Müller"),
  email = whoami::email_address(fallback = "mail@example.com"),
  date = "2020-12-12",
  dir = file.path(tempdir(), "fledge"),
  news = FALSE
)
}
\arguments{
\item{open}{Whether to open the new project.}

\item{name}{Package name.}

\item{maintainer}{Name for DESCRIPTION and git.}

\item{email}{Email for DESCRIPTION and git.}

\item{date}{String of time for DESCRIPTION and git.}

\item{dir}{Directory within which to create the mock package folder.}

\item{news}{If TRUE, create a NEWS.md file.}
}
\value{
The path to the newly created mock package.
}
\description{
Create example repo for fledge demos
}
