% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-get-last-tag.R
\name{get_last_tag}
\alias{get_last_tag}
\title{The most recent tag}
\usage{
get_last_tag()
}
\value{
An object of the repo's \code{\link[gert:git_tag]{gert::git_tag_list()}}.
}
\description{
Returns the most recent Git tag.
}
