% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-get-last-tag.R
\name{get_last_tag}
\alias{get_last_tag}
\title{The most recent tag}
\usage{
get_last_tag()
}
\value{
The SHA of the commit representing the last tag.
For annotated tags (as created by fledge), the SHA may be different
from the SHA of the commit that this tag points to.
Use \code{\link[gert:git_commit]{gert::git_log()}} to find the actual commit.
}
\description{
Returns the most recent Git tag.
}
