% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-get-top-level-commits.R
\name{get_top_level_commits}
\alias{get_top_level_commits}
\title{All top-level commits}
\usage{
get_top_level_commits(since)
}
\arguments{
\item{since}{A commit SHA, e.g. as returned from \code{\link[=get_last_tag]{get_last_tag()}}.}
}
\value{
A character vector of commit SHAs.
}
\description{
Return all top-level commits since a particular version
as commit objects.
}
