% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{release}
\alias{release}
\alias{pre_release}
\alias{post_release}
\title{Automating CRAN release}
\usage{
pre_release(which = "patch")

release()

post_release()
}
\description{
\code{pre_release()} is run when development of a package is finished
and it is ready to be sent to CRAN.

\code{release()} sends to CRAN after performing several checks,
and offers help with accepting the submission.

\code{post_release()} should be called after the submission has been accepted.
}
