% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{release}
\alias{release}
\alias{pre_release}
\alias{post_release}
\title{Automating CRAN release}
\usage{
pre_release(which = "patch", force = FALSE)

release()

post_release()
}
\arguments{
\item{which}{Component of the version number to update. Supported
values are
\itemize{
\item \code{"patch"} (default),
\item \code{"minor"},
\item \code{"major"}.
}}

\item{force}{Create branches and tags even if they exist.
Useful to recover from a previously failed attempt.}
}
\description{
\code{pre_release()} is run when development of a package is finished
and it is ready to be sent to CRAN.

\code{release()} sends to CRAN after performing several checks,
and offers help with accepting the submission.

\code{post_release()} should be called after the submission has been accepted.
}
\details{
\code{pre_release()}:
\itemize{
\item Ensures that no modified files are in the git index.
\item Bumps version to a non-development version which should be sent to CRAN.
\item Writes/updates \code{cran-comments.md} with useful information about the current
release process.
\item Prompts the user to run \code{devtools::check_win_devel()}.
\item Prompts the user to run \code{rhub::check_for_cran()}.
}
}
