% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-unbump-version.R, R/unbump-version.R
\name{unbump_version}
\alias{unbump_version}
\alias{unbump_version_impl}
\title{Undoes bumping the package version}
\usage{
unbump_version()
}
\description{
This undoes the effect of a \code{\link[=bump_version]{bump_version()}} call, with a safety check.
}
