% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-update-news.R
\name{update_news}
\alias{update_news}
\title{Update NEWS.md with messages from top-level commits}
\usage{
update_news(range = get_top_level_commits(since = get_last_tag()))
}
\arguments{
\item{range}{A commit range as character vector of SHAs,
e.g. as returned from \code{\link[=get_top_level_commits]{get_top_level_commits()}}.}
}
\description{
Lists all commits from a range (default: top-level commits since the most
recent tag) and adds bullets from their body to \code{NEWS.md}.
}
