% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-update-news.R
\name{update_news}
\alias{update_news}
\title{Update NEWS.md with messages from top-level commits}
\usage{
update_news(messages = NULL)
}
\arguments{
\item{messages}{A character vector of commit messages,
e.g. as in the \code{message} column in the return value of \code{\link[=get_top_level_commits]{get_top_level_commits()}}.
The default uses the top level commits since the last tag as retrieved by \code{\link[=get_last_tag]{get_last_tag()}}.}
}
\description{
Lists all commits from a range (default: top-level commits since the most
recent tag) and adds bullets from their body to \code{NEWS.md}.
}
