% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-update-version.R
\name{update_version}
\alias{update_version}
\title{Update NEWS.md and DESCRIPTION with a new version}
\usage{
update_version(which = "dev")
}
\arguments{
\item{which}{Component of the version number to update. Supported
values are \code{"dev"} (default), \code{"patch"}, \code{"minor"} and \code{"major"}.}
}
\description{
Bumps a version component and adds to \code{NEWS.md} and \code{DESCRIPTION}.
}
